Remove ENABLED/DISABLED log messages and add a comment explaining why
Deficiencies in the gnome-shell extensions framework make it
impossible to implement this extension "properly".

The problem is that there is no way to distinguish in `disable()`
between when it is called because the extension was disabled or
unloaded (e.g., on logout) vs. when the screen is being
locked:(apparently, gnome-shell disables all extensions when the
screen is locked!).

The extension should not revert the `inhibit_remote_access()` function
when the screen is locked, but it has to, because it can't tell that
`disable()` is just being called because the screen is being locked.

At least for the time being, the extension still _works_, because
`inhibit_remote_access()` gets called before the screen is locked,
i.e., before `disable()` gets called and it gets reverted. But this is
undocumented behavior which could change at any time.

Furthermore, while I would like to be able to log messages telling the
user when remote desktop access while the screen is locked is enabled
vs. disabled, because of the issue described above I can't actually do
that, at least not without leaving `inhibit_remote_access()` modified
even after disable() is called, which I'm not allowed to do. This is
why I'm removing these log messages in this commit: they're useless
and misleading.

Perhaps at some point the maintainers of gnome-shell will add
functionality giving extensions more visibility into what is going on
vis a vis install, uninstall, enable, disable, lock screen, etc., but
for the time being the best we can hope for is that the undocumented
behavior of when `inhibit_remote_access()` is called doesn't change.